@@ -152,27 +120,26 @@
     "@angular/compiler-cli": "^16.2.5",
     "@compodoc/compodoc": "^1.1.15",
     "@cypress/schematic": "^2.5.1",
+    "@swc/core": "1.3.82",
+    "@swc/wasm": "1.3.82",
     "@trivago/prettier-plugin-sort-imports": "^4.2.0",
     "@types/flexsearch": "^0.7.3",
     "@types/geojson": "^7946.0.10",
     "@types/hammerjs": "^2.0.36",
-    "@types/html2canvas": "1.0.0",
     "@types/jasmine": "~4.3.0",
     "@types/lodash": "^4.14.195",
     "@types/node": "^20.6.2",
     "@typescript-eslint/eslint-plugin": "^6.2.0",
     "@typescript-eslint/parser": "^6.2.0",
+    "@typescript-eslint/types": "^6.2.0",
     "angular-cli-ghpages": "^1.0.3",
     "conventional-changelog-cli": "^4.1.0",
     "copy-newer": "^2.1.2",
     "cypress": "^13.2.0",
     "del": "^7.0.0",
-    "eslint": "^8.46.0",
+    "eslint": "^8.47.0",
     "eslint-plugin-cypress": "^2.14.0",
     "eslint-plugin-unused-imports": "^3.0.0",
-    "gulp": "^4.0.0",
-    "gulp-json-editor": "^2.5.0",
-    "gulp-replace": "^1.1.0",
     "jasmine-core": "~5.1.0",
     "jasmine-spec-reporter": "~7.0.0",
     "karma": "~6.4.2",
