@@ -6,6 +6,8 @@ import {
   Validators
 } from '@angular/forms';
 
+import { MediaService } from '@igo2/core';
+
 import { BehaviorSubject } from 'rxjs';
 
 import { PrintOptions } from '../shared/print.interface';
@@ -28,7 +30,7 @@ export class PrintFormComponent implements OnInit {
   public outputFormats = PrintOutputFormat;
   public paperFormats = PrintPaperFormat;
   public orientations = PrintOrientation;
-  public resolutions = PrintResolution;
+  public resolutions = [...PrintResolution];
   public imageFormats = PrintSaveImageFormat;
   public legendPositions = PrintLegendPosition;
   public isPrintService = true;
@@ -80,7 +82,7 @@ export class PrintFormComponent implements OnInit {
     return this.resolutionField.value;
   }
   set resolution(value: PrintResolution) {
-    this.resolutionField.setValue(value || PrintResolution['96'], {
+    this.resolutionField.setValue(value || ('96' satisfies PrintResolution), {
       onlySelf: true
     });
   }
@@ -202,7 +204,10 @@ export class PrintFormComponent implements OnInit {
 
   maxLength: number = 180;
 
-  constructor(private formBuilder: UntypedFormBuilder) {
+  constructor(
+    private formBuilder: UntypedFormBuilder,
+    private mediaService: MediaService
+  ) {
     this.form = this.formBuilder.group({
       title: ['', [Validators.minLength(0), Validators.maxLength(130)]],
       subtitle: ['', [Validators.minLength(0), Validators.maxLength(120)]],
@@ -225,6 +230,11 @@ export class PrintFormComponent implements OnInit {
 
   ngOnInit() {
     this.doZipFileField.setValue(false);
+    if (this.mediaService.isMobile()) {
+      this.resolutions = this.resolutions.filter(
+        (resolution) => resolution !== '300'
+      );
+    }
   }
 
   handleFormSubmit(data: PrintOptions, isValid: boolean) {
