@@ -19,8 +19,8 @@ export type PrintPaperFormat = keyof typeof PrintPaperFormat;
 export const PrintOrientation = strEnum(['landscape', 'portrait']);
 export type PrintOrientation = keyof typeof PrintOrientation;
 
-export const PrintResolution = strEnum(['72', '96', '150']); // For now, we remove the 300 dpi because there is too much instability on iOS and slow device.
-export type PrintResolution = keyof typeof PrintResolution;
+export const PrintResolution = ['72', '96', '150', '300'] as const;
+export type PrintResolution = (typeof PrintResolution)[number];
 
 export const PrintSaveImageFormat = strEnum([
   'Bmp',
