@@ -1,6 +1,7 @@
 import { compile } from 'sass';
 
 import { createFile } from './file-system.utils';
+import { PATHS } from '../config/paths';
 
 export async function compileStyle(
   input: string,
@@ -8,7 +9,7 @@ export async function compileStyle(
   outputFileName: string
 ) {
   const result = compile(input, {
-    loadPaths: ['node_modules']
+    loadPaths: [PATHS.nodeModules]
   });
   await createFile(outputFileName, destination, result.css);
 }
